+++ spec/presence.js
@@ -144,10 +144,10 @@
             '    <delay xmlns="urn:xmpp:delay" stamp="2017-02-15T20:26:05Z" from="'+contact_jid+'/priority-1-resource"/>'+
             '</presence>');
             _converse.connection._dataRecv(test_utils.createRequest(stanza[0]));
-            expect(contact.get('chat_status')).toBe('online');
-            expect(_.keys(contact.get('resources')).length).toBe(1);
-            expect(contact.get('resources')['priority-1-resource']['priority']).toBe(1);
-            expect(contact.get('resources')['priority-1-resource']['status']).toBe('online');
+            expect(contact.presence.get('show')).toBe('online');
+            expect(_.keys(contact.presence.get('resources')).length).toBe(1);
+            expect(contact.presence.get('resources')['priority-1-resource']['priority']).toBe(1);
+            expect(contact.presence.get('resources')['priority-1-resource']['show']).toBe('online');
 
             stanza = $(
             '<presence xmlns="jabber:client"'+
@@ -161,12 +161,12 @@
             '    <delay xmlns="urn:xmpp:delay" stamp="2017-02-15T17:02:24Z" from="'+contact_jid+'/priority-0-resource"/>'+
             '</presence>');
             _converse.connection._dataRecv(test_utils.createRequest(stanza[0]));
-            expect(_converse.roster.get(contact_jid).get('chat_status')).toBe('online');
-            expect(_.keys(contact.get('resources')).length).toBe(2);
-            expect(contact.get('resources')['priority-0-resource']['priority']).toBe(0);
-            expect(contact.get('resources')['priority-0-resource']['status']).toBe('xa');
-            expect(contact.get('resources')['priority-1-resource']['priority']).toBe(1);
-            expect(contact.get('resources')['priority-1-resource']['status']).toBe('online');
+            expect(_converse.roster.get(contact_jid).presence.get('show')).toBe('online');
+            expect(_.keys(contact.presence.get('resources')).length).toBe(2);
+            expect(contact.presence.get('resources')['priority-0-resource']['priority']).toBe(0);
+            expect(contact.presence.get('resources')['priority-0-resource']['show']).toBe('xa');
+            expect(contact.presence.get('resources')['priority-1-resource']['priority']).toBe(1);
+            expect(contact.presence.get('resources')['priority-1-resource']['show']).toBe('online');
 
             stanza = $(
             '<presence xmlns="jabber:client"'+
@@ -176,14 +176,14 @@
             '    <show>dnd</show>'+
             '</presence>');
             _converse.connection._dataRecv(test_utils.createRequest(stanza[0]));
-            expect(_converse.roster.get(contact_jid).get('chat_status')).toBe('dnd');
-            expect(_.keys(contact.get('resources')).length).toBe(3);
-            expect(contact.get('resources')['priority-0-resource']['priority']).toBe(0);
-            expect(contact.get('resources')['priority-0-resource']['status']).toBe('xa');
-            expect(contact.get('resources')['priority-1-resource']['priority']).toBe(1);
-            expect(contact.get('resources')['priority-1-resource']['status']).toBe('online');
-            expect(contact.get('resources')['priority-2-resource']['priority']).toBe(2);
-            expect(contact.get('resources')['priority-2-resource']['status']).toBe('dnd');
+            expect(_converse.roster.get(contact_jid).presence.get('show')).toBe('dnd');
+            expect(_.keys(contact.presence.get('resources')).length).toBe(3);
+            expect(contact.presence.get('resources')['priority-0-resource']['priority']).toBe(0);
+            expect(contact.presence.get('resources')['priority-0-resource']['show']).toBe('xa');
+            expect(contact.presence.get('resources')['priority-1-resource']['priority']).toBe(1);
+            expect(contact.presence.get('resources')['priority-1-resource']['show']).toBe('online');
+            expect(contact.presence.get('resources')['priority-2-resource']['priority']).toBe(2);
+            expect(contact.presence.get('resources')['priority-2-resource']['show']).toBe('dnd');
 
             stanza = $(
             '<presence xmlns="jabber:client"'+
@@ -193,16 +193,16 @@
             '    <show>away</show>'+
             '</presence>');
             _converse.connection._dataRecv(test_utils.createRequest(stanza[0]));
-            expect(_converse.roster.get(contact_jid).get('chat_status')).toBe('away');
-            expect(_.keys(contact.get('resources')).length).toBe(4);
-            expect(contact.get('resources')['priority-0-resource']['priority']).toBe(0);
-            expect(contact.get('resources')['priority-0-resource']['status']).toBe('xa');
-            expect(contact.get('resources')['priority-1-resource']['priority']).toBe(1);
-            expect(contact.get('resources')['priority-1-resource']['status']).toBe('online');
-            expect(contact.get('resources')['priority-2-resource']['priority']).toBe(2);
-            expect(contact.get('resources')['priority-2-resource']['status']).toBe('dnd');
-            expect(contact.get('resources')['priority-3-resource']['priority']).toBe(3);
-            expect(contact.get('resources')['priority-3-resource']['status']).toBe('away');
+            expect(_converse.roster.get(contact_jid).presence.get('show')).toBe('away');
+            expect(_.keys(contact.presence.get('resources')).length).toBe(4);
+            expect(contact.presence.get('resources')['priority-0-resource']['priority']).toBe(0);
+            expect(contact.presence.get('resources')['priority-0-resource']['show']).toBe('xa');
+            expect(contact.presence.get('resources')['priority-1-resource']['priority']).toBe(1);
+            expect(contact.presence.get('resources')['priority-1-resource']['show']).toBe('online');
+            expect(contact.presence.get('resources')['priority-2-resource']['priority']).toBe(2);
+            expect(contact.presence.get('resources')['priority-2-resource']['show']).toBe('dnd');
+            expect(contact.presence.get('resources')['priority-3-resource']['priority']).toBe(3);
+            expect(contact.presence.get('resources')['priority-3-resource']['show']).toBe('away');
 
             stanza = $(
             '<presence xmlns="jabber:client"'+
@@ -213,18 +213,18 @@
             '    <delay xmlns="urn:xmpp:delay" stamp="2017-02-15T15:02:24Z" from="'+contact_jid+'/older-priority-1-resource"/>'+
             '</presence>');
             _converse.connection._dataRecv(test_utils.createRequest(stanza[0]));
-            expect(_converse.roster.get(contact_jid).get('chat_status')).toBe('away');
-            expect(_.keys(contact.get('resources')).length).toBe(5);
-            expect(contact.get('resources')['older-priority-1-resource']['priority']).toBe(1);
-            expect(contact.get('resources')['older-priority-1-resource']['status']).toBe('dnd');
-            expect(contact.get('resources')['priority-0-resource']['priority']).toBe(0);
-            expect(contact.get('resources')['priority-0-resource']['status']).toBe('xa');
-            expect(contact.get('resources')['priority-1-resource']['priority']).toBe(1);
-            expect(contact.get('resources')['priority-1-resource']['status']).toBe('online');
-            expect(contact.get('resources')['priority-2-resource']['priority']).toBe(2);
-            expect(contact.get('resources')['priority-2-resource']['status']).toBe('dnd');
-            expect(contact.get('resources')['priority-3-resource']['priority']).toBe(3);
-            expect(contact.get('resources')['priority-3-resource']['status']).toBe('away');
+            expect(_converse.roster.get(contact_jid).presence.get('show')).toBe('away');
+            expect(_.keys(contact.presence.get('resources')).length).toBe(5);
+            expect(contact.presence.get('resources')['older-priority-1-resource']['priority']).toBe(1);
+            expect(contact.presence.get('resources')['older-priority-1-resource']['show']).toBe('dnd');
+            expect(contact.presence.get('resources')['priority-0-resource']['priority']).toBe(0);
+            expect(contact.presence.get('resources')['priority-0-resource']['show']).toBe('xa');
+            expect(contact.presence.get('resources')['priority-1-resource']['priority']).toBe(1);
+            expect(contact.presence.get('resources')['priority-1-resource']['show']).toBe('online');
+            expect(contact.presence.get('resources')['priority-2-resource']['priority']).toBe(2);
+            expect(contact.presence.get('resources')['priority-2-resource']['show']).toBe('dnd');
+            expect(contact.presence.get('resources')['priority-3-resource']['priority']).toBe(3);
+            expect(contact.presence.get('resources')['priority-3-resource']['show']).toBe('away');
 
             stanza = $(
             '<presence xmlns="jabber:client"'+
@@ -233,16 +233,16 @@
             '          from="'+contact_jid+'/priority-3-resource">'+
             '</presence>');
             _converse.connection._dataRecv(test_utils.createRequest(stanza[0]));
-            expect(_converse.roster.get(contact_jid).get('chat_status')).toBe('dnd');
-            expect(_.keys(contact.get('resources')).length).toBe(4);
-            expect(contact.get('resources')['priority-0-resource']['priority']).toBe(0);
-            expect(contact.get('resources')['priority-0-resource']['status']).toBe('xa');
-            expect(contact.get('resources')['priority-1-resource']['priority']).toBe(1);
-            expect(contact.get('resources')['priority-1-resource']['status']).toBe('online');
-            expect(contact.get('resources')['priority-2-resource']['priority']).toBe(2);
-            expect(contact.get('resources')['priority-2-resource']['status']).toBe('dnd');
-            expect(contact.get('resources')['older-priority-1-resource']['priority']).toBe(1);
-            expect(contact.get('resources')['older-priority-1-resource']['status']).toBe('dnd');
+            expect(_converse.roster.get(contact_jid).presence.get('show')).toBe('dnd');
+            expect(_.keys(contact.presence.get('resources')).length).toBe(4);
+            expect(contact.presence.get('resources')['priority-0-resource']['priority']).toBe(0);
+            expect(contact.presence.get('resources')['priority-0-resource']['show']).toBe('xa');
+            expect(contact.presence.get('resources')['priority-1-resource']['priority']).toBe(1);
+            expect(contact.presence.get('resources')['priority-1-resource']['show']).toBe('online');
+            expect(contact.presence.get('resources')['priority-2-resource']['priority']).toBe(2);
+            expect(contact.presence.get('resources')['priority-2-resource']['show']).toBe('dnd');
+            expect(contact.presence.get('resources')['older-priority-1-resource']['priority']).toBe(1);
+            expect(contact.presence.get('resources')['older-priority-1-resource']['show']).toBe('dnd');
 
             stanza = $(
             '<presence xmlns="jabber:client"'+
@@ -251,14 +251,14 @@
             '          from="'+contact_jid+'/priority-2-resource">'+
             '</presence>');
             _converse.connection._dataRecv(test_utils.createRequest(stanza[0]));
-            expect(_converse.roster.get(contact_jid).get('chat_status')).toBe('online');
-            expect(_.keys(contact.get('resources')).length).toBe(3);
-            expect(contact.get('resources')['priority-0-resource']['priority']).toBe(0);
-            expect(contact.get('resources')['priority-0-resource']['status']).toBe('xa');
-            expect(contact.get('resources')['priority-1-resource']['priority']).toBe(1);
-            expect(contact.get('resources')['priority-1-resource']['status']).toBe('online');
-            expect(contact.get('resources')['older-priority-1-resource']['priority']).toBe(1);
-            expect(contact.get('resources')['older-priority-1-resource']['status']).toBe('dnd');
+            expect(_converse.roster.get(contact_jid).presence.get('show')).toBe('online');
+            expect(_.keys(contact.presence.get('resources')).length).toBe(3);
+            expect(contact.presence.get('resources')['priority-0-resource']['priority']).toBe(0);
+            expect(contact.presence.get('resources')['priority-0-resource']['show']).toBe('xa');
+            expect(contact.presence.get('resources')['priority-1-resource']['priority']).toBe(1);
+            expect(contact.presence.get('resources')['priority-1-resource']['show']).toBe('online');
+            expect(contact.presence.get('resources')['older-priority-1-resource']['priority']).toBe(1);
+            expect(contact.presence.get('resources')['older-priority-1-resource']['show']).toBe('dnd');
 
             stanza = $(
             '<presence xmlns="jabber:client"'+
@@ -267,12 +267,12 @@
             '          from="'+contact_jid+'/priority-1-resource">'+
             '</presence>');
             _converse.connection._dataRecv(test_utils.createRequest(stanza[0]));
-            expect(_converse.roster.get(contact_jid).get('chat_status')).toBe('dnd');
-            expect(_.keys(contact.get('resources')).length).toBe(2);
-            expect(contact.get('resources')['priority-0-resource']['priority']).toBe(0);
-            expect(contact.get('resources')['priority-0-resource']['status']).toBe('xa');
-            expect(contact.get('resources')['older-priority-1-resource']['priority']).toBe(1);
-            expect(contact.get('resources')['older-priority-1-resource']['status']).toBe('dnd');
+            expect(_converse.roster.get(contact_jid).presence.get('show')).toBe('dnd');
+            expect(_.keys(contact.presence.get('resources')).length).toBe(2);
+            expect(contact.presence.get('resources')['priority-0-resource']['priority']).toBe(0);
+            expect(contact.presence.get('resources')['priority-0-resource']['show']).toBe('xa');
+            expect(contact.presence.get('resources')['older-priority-1-resource']['priority']).toBe(1);
+            expect(contact.presence.get('resources')['older-priority-1-resource']['show']).toBe('dnd');
 
             stanza = $(
             '<presence xmlns="jabber:client"'+
@@ -281,10 +281,10 @@
             '          from="'+contact_jid+'/older-priority-1-resource">'+
             '</presence>');
             _converse.connection._dataRecv(test_utils.createRequest(stanza[0]));
-            expect(_converse.roster.get(contact_jid).get('chat_status')).toBe('xa');
-            expect(_.keys(contact.get('resources')).length).toBe(1);
-            expect(contact.get('resources')['priority-0-resource']['priority']).toBe(0);
-            expect(contact.get('resources')['priority-0-resource']['status']).toBe('xa');
+            expect(_converse.roster.get(contact_jid).presence.get('show')).toBe('xa');
+            expect(_.keys(contact.presence.get('resources')).length).toBe(1);
+            expect(contact.presence.get('resources')['priority-0-resource']['priority']).toBe(0);
+            expect(contact.presence.get('resources')['priority-0-resource']['show']).toBe('xa');
 
             stanza = $(
             '<presence xmlns="jabber:client"'+
@@ -293,8 +293,8 @@
             '          from="'+contact_jid+'/priority-0-resource">'+
             '</presence>');
             _converse.connection._dataRecv(test_utils.createRequest(stanza[0]));
-            expect(_converse.roster.get(contact_jid).get('chat_status')).toBe('offline');
-            expect(_.keys(contact.get('resources')).length).toBe(0);
+            expect(_converse.roster.get(contact_jid).presence.get('show')).toBe('offline');
+            expect(_.keys(contact.presence.get('resources')).length).toBe(0);
             done();
         }));
     });